Move types to their own files
